Create components folder and move App.css and App.js
